Split into lib and bin
